tools/analytics/requirements.txt
google-analytics-data>=0.18.0
google-auth>=2.0.0
google-api-python-client>=2.0.0
